refactor(@galactiks/explorer): move repositories in a dedicated directory and add get per type one
