+++ packages/explorer/src/core/content/index.mts
@@ -1,2 +1,2 @@
-export * from './repository.mjs';
+export * from './repositories/index.mjs';
 export * from './types/index.mjs';
