+++ packages/explorer/src/core/content/repositories/filters.mts
@@ -1,0 +1,10 @@
+import type { Content } from '../types/index.mjs';
+
+export type RepositoryFilters = {
+  type?: Content['type'];
+  inLanguages?: string[];
+};
+
+export type WebPageElementFilters = {
+  inLanguage?: string;
+};
